requirements.txt
Django==5.1
djangorestframework==3.15.2
psycopg==3.2.1
celery==5.4.0
redis==5.0.8
shortuuid==1.0.13
python-dotenv==1.0.1
whitenoise[brotli]==6.7.0
pillow==10.4.0
gunicorn==23.0.0
scout-apm==3.1.0
sentry-sdk[django]==2.13.0
dj-database-url==2.2.0
django-vite==3.0.4
pytest-playwright==0.5.2
